+++ packages/docusaurus-types/src/index.d.ts
@@ -30,7 +30,7 @@ export interface DocusaurusConfig {
   tagline?: string;
   title: string;
   url: string;
-  // trailingSlash undefined = legacy retrocompatible behavior => /file => /fileName/index.html
+  // trailingSlash undefined = legacy retrocompatible behavior => /file => /file/index.html
   trailingSlash: boolean | undefined;
   i18n: I18nConfig;
   onBrokenLinks: ReportingSeverity;
