+++ packages/docusaurus/src/client/exports/applyTrailingSlash.tsx
@@ -1,0 +1,29 @@
+/**
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+export default function applyTrailingSlash(
+  path: string,
+  trailingSlash: boolean | undefined,
+): string {
+  function addTrailingSlash(str: string): string {
+    return str.endsWith('/') ? str : `${str}/`;
+  }
+  function removeTrailingSlash(str: string): string {
+    return str.endsWith('/') ? str.slice(0, -1) : str;
+  }
+  // undefined = legacy retrocompatible behavior
+  if (typeof trailingSlash === 'undefined') {
+    return path;
+  }
+
+  // The trailing slash should be handled before the querystring/hash !
+  const [pathname] = path.split(/[#?]/);
+  const newPathname = trailingSlash
+    ? addTrailingSlash(pathname)
+    : removeTrailingSlash(pathname);
+  return path.replace(pathname, newPathname);
+}
